Cygwin compat.

+++ hcore/hsocket.cxx
@@ -241,7 +241,7 @@ HSocket::ptr_t HSocket::accept( void ) {
 	socket->_addressSize = static_cast<int>( addressSize );
 	socket->_needShutdown = true;
 	socket->set_timeout( _timeout );
-	::memcpy( socket->_address, address, addressSize );
+	::memcpy( socket->_address, address, static_cast<size_t>( addressSize ) );
 	if ( _clients->find( fileDescriptor ) != _clients->end() )
 		M_THROW( _( "inconsitent client list state" ), fileDescriptor );
 	_clients->insert( make_pair( fileDescriptor, socket ) );
